backend/pyproject.toml
[project]
name = "medical-chatbot-backend"
version = "0.1.0"
description = "LLM + RAGraph Medical Chatbot Backend"
authors = [{name = "Your Name"}]
requires-python = ">=3.9"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "neo4j",
    "pandas",
    "numpy",
    "sentence-transformers",
    "streamlit",
    "python-dotenv",
    "groq>=0.31.1",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    # Testing tools
    "pytest",
    "pytest-cov"
]

[tool.uv]
# optional: configure caching, index settings here
